# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MarketRisk = typeof MarketRisk[keyof typeof MarketRisk];


export const MarketRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Market {
  pick: string;
  type: string;
  odds: number;
  open_odds: number;
  sharp_odds: number;
  score: number;
  risk: MarketRisk;
  reasons: string[];
  ev?: number;
}

export interface Game {
  sport: string;
  league: string;
  home: string;
  away: string;
  starts_at: string;
  start_in_minutes: number;
  markets: Market[];
}

export interface GamesResponse {
  count: number;
  games: Game[];
}

export interface Pick {
  pick: string;
  type: string;
  odds: number;
  open_odds: number;
  sharp_odds: number;
  score: number;
  risk: string;
  reasons: string[];
  ev?: number;
  sport: string;
  league: string;
  home: string;
  away: string;
  game_key: string;
  start_in_minutes: number;
}

export interface Combo {
  type: string;
  total_odds: number;
  avg_score: number;
  picks: Pick[];
}

export interface RecommendationsResponse {
  sport: string;
  window: number;
  games: Game[];
  recommendations: Combo[];
  excluded: Pick[];
  mode: string;
  notice: string;
}

export type GetGamesParams = {
sport?: GetGamesSport;
window?: number;
};

export type GetGamesSport = typeof GetGamesSport[keyof typeof GetGamesSport];


export const GetGamesSport = {
  all: 'all',
  soccer: 'soccer',
  baseball: 'baseball',
} as const;

export type GetRecommendationsParams = {
sport?: GetRecommendationsSport;
window?: number;
};

export type GetRecommendationsSport = typeof GetRecommendationsSport[keyof typeof GetRecommendationsSport];


export const GetRecommendationsSport = {
  all: 'all',
  soccer: 'soccer',
  baseball: 'baseball',
} as const;