# lib/api-zod/src/generated/types/combo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Pick } from './pick';

export interface Combo {
  type: string;
  total_odds: number;
  avg_score: number;
  picks: Pick[];
}