# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Market } from './market';

export interface Game {
  sport: string;
  league: string;
  home: string;
  away: string;
  starts_at: string;
  start_in_minutes: number;
  markets: Market[];
}