# lib/api-zod/src/generated/types/gamesResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from './game';

export interface GamesResponse {
  count: number;
  games: Game[];
}