# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketRisk } from './marketRisk';

export interface Market {
  pick: string;
  type: string;
  odds: number;
  open_odds: number;
  sharp_odds: number;
  score: number;
  risk: MarketRisk;
  reasons: string[];
  ev?: number;
}