# lib/api-zod/src/generated/types/pick.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Pick {
  pick: string;
  type: string;
  odds: number;
  open_odds: number;
  sharp_odds: number;
  score: number;
  risk: string;
  reasons: string[];
  ev?: number;
  sport: string;
  league: string;
  home: string;
  away: string;
  game_key: string;
  start_in_minutes: number;
}