# lib/api-zod/src/generated/types/recommendationsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Combo } from './combo';
import type { Game } from './game';
import type { Pick } from './pick';

export interface RecommendationsResponse {
  sport: string;
  window: number;
  games: Game[];
  recommendations: Combo[];
  excluded: Pick[];
  mode: string;
  notice: string;
}